requirements.txt
argon2-cffi==25.1.0
backports.tarfile==1.2.0
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
django-allauth==65.11.2
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-coverage-plugin==3.1.1
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-extensions==4.1
django-model-utils==5.0.0
django-redis==6.0.0
django-stubs==5.2.5
djangorestframework-simplejwt==5.5.1
djlint==1.36.4
drf-spectacular==0.28.0
factory-boy==3.3.3
fido2==2.0.0
flower==2.0.1
gunicorn==23.0.0
hiredis==3.2.1
importlib-metadata==8.0.0
ipdb==0.13.13
jaraco.collections==5.1.0
mypy==1.18.2
pillow==11.3.0
pip-chill==1.0.3
pre-commit==4.3.0
psycopg==3.2.10
psycopg-c==3.2.10
pytest-django==4.11.1
pytest-sugar==1.1.1
python-decouple==3.8
python-slugify==8.0.4
qrcode==8.2
ruff==0.13.3
sphinx-autobuild==2025.8.25
tomli==2.0.1
watchdog==6.0.0
werkzeug==3.1.3
whitenoise==6.11.0



{
  "promotion": "cca0b5da-9d89-48fd-bef8-2dadd375c07f",
  "user": "22b06af8-365b-45c6-b3f5-6286cd3e8996",
  "order": "ebb8a6c6-eb19-4709-915a-f29c43f6deab"
  
}


{
  "minimum_purchase_amount": "100.00",
  "minimum_quantity": 1
}
